Compile a version string into the CLVTreeScaper binary for every build. This string contains the latest release, the commit hash if commits have been made since the release, and dirty if there are any uncommitted modifications to tracked files
Additionally, the -v and -version flags have been added to output the
version.

This also updates the Makefile to simply rebuild the binary each time.
Previously the binary was not rebuilt, even if source files were
modified. Because the TreeScaper rewrite project includes a new Makefile, we
can add further dependency handling there.