Add ability to run tree inference remotely via the CIPRES REST API (CRA).
Specifically, RAxML on XSEDE HPC resources is used. The current
functionality is asynchronous in that the -inference flag
it used to submit a job, and once the job is complete, TreeScaper can be run
again with the -trees flag and the job name to download the bootstrap
trees and run tree analysis.

The user is currently responsible for providing their own username,
password, and application ID.