Adds support for multiple tree inference jobs via CRA.
A file is provided with -f that contains list of input sequence files.
A new RAxML job is submitted for each sequence files. The jobs are
monitored for completion, and their results downloaded in a "cra_output"
directory once finished.

An internal queue is utilized to submit no more than a given number of
jobs at a time. The CRA enforces an ACTIVE_LIMIT via the API, and so
this limit is reflected within TreeScaper.

Once the limit is reached, new jobs are not submitted until active jobs
have finished.

The following rate limits are imposed in TreeScaper code:
- No more than 1 REST request per 0.5 seconds.
- No more than 1 job status request per job per 10 seconds.

The option to use the -tree option with a CRA job name is removed. All
CRA job output should now be downloaded as part of the -inference
sequence.

A file "cra_output/job_status" tracks the status of jobs. In the future
this file may be used to resume interrupted sessions.