Compare taxa in tree string with the taxon list.
The taxon list may be incomplete. If new taxon is found, it will be inserted into the list. Whether there is a new taxon found in tree string or an absent taxon from the list not found in the tree string, a warning message will be generated and the routine is expected to be terminated with a TaxonList_postfix.out for details leaf set info.