Scan the leaf set
This routine first check every tree string with the existing taxon list. During the process, the taxon list will collect all taxa encountered and therefore it is assumed to be the complete set. Then if non-identical leaf sets are found from the tree strings, the routine produce a file that output all taxa encountered as well as the set of absent taxa in every tree string.